requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI and ML dependencies for question generation and RAG
ollama==0.3.3
sentence-transformers==2.2.2
PyPDF2==3.0.1
numpy==1.24.3
torch==2.0.1
transformers==4.35.2

# LangChain dependencies for vector store optimization
langchain-community
langchain-ollama
faiss-cpu
pypdf
tiktoken

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2
